+++ CHANGELOG.md
@@ -14,7 +14,9 @@ Version 6.11
 * Modified `gams-view-lst` and `gams-jump-to-lst`. If you attch the
   universal-argument (C-u) to these commands, you can show the LST file
   buffer next to the GAMS buffer.
-
+  
+* Added the new function `gams-fill-paragraph` and bind it to "M-q" in
+  GAMS mode.
 
 
 Version 6.10